# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KPI Emanuel Corp API",
        "description": "API Documentation untuk Aplikasi KPI Emanuel Corp",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/form-submissions/{id}": {
            "get": {
                "tags": [
                    "Form Builder"
                ],
                "summary": "Get detail submission form",
                "operationId": "7edc2d9bb32e503828fceffd9643a3d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail submission form"
                    },
                    "404": {
                        "description": "Submission tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Form Builder"
                ],
                "summary": "Hapus submission (owner only)",
                "operationId": "b6a6f67962ccdbe120a87badc3884c44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission berhasil dihapus"
                    },
                    "404": {
                        "description": "Submission tidak ditemukan"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/form-submissions/form/{formTemplateId}": {
            "get": {
                "tags": [
                    "Form Builder"
                ],
                "summary": "Get submission berdasarkan form template",
                "operationId": "71b8e05b41c5113d93e7e345b6613065",
                "parameters": [
                    {
                        "name": "formTemplateId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List submission by form template"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}